Update 1.5.1 Basic string processing skills.cpp
Given a text file that contains only alphabet characters [A-Za-z], digits [0-9], spaces, and periods ('.'), write a program to read this text file line by line until we encounter a line that starts with seven periods ("......."). Concatenate (combine) each line into one long string T. When two lines are combined, give one space between them so that the last word of the previous line is separated from the first word of the current line. There can be up to 30 characters per line and no more than 10 lines for this input block. There are no trailing spaces at the end of each line and each line ends with a newline character. Note that the sample input is shown inside a box after question