Create 1.5.2 Basic string processing skills.cpp
Suppose that we have one long string T. We want to check if another string P can be found in T. Report all the indices where P appears in T or report -1 if P cannot be found in T. For example, if T = "I love CS3233 Competitive Programming. i also love AlGoRiThM" and P = "I", then the output is only {0} (0-based indexing) because uppercase 'I' and lowercase 'i' are considered different and thus the character 'i' at index {39} is not part of the output. If P = "love", then the output is {2, 46). If P = "book", then the output is {-1}.